Added Admin Parking edit

+++ src/main/java/com/example/application/data/service/ParkingSlotRepository.java
@@ -1,9 +1,12 @@
 package com.example.application.data.service;
 
 import com.example.application.data.entity.ParkingSlot;
+import com.example.application.data.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParkingSlotRepository extends JpaRepository<ParkingSlot, Long> {
+import java.util.Optional;
+
+public interface ParkingSlotRepository extends JpaRepository<ParkingSlot, Integer> {
 
     ParkingSlot getByName(String name);
 }
